+++ app/webpacker/components/RegistrationsV2/Registrations/PsychSheetEventSelector.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import _ from 'lodash';
+import { EventSelector } from '../../wca/EventSelector';
+
+export default function PsychSheetEventSelector({
+  handleEventSelection,
+  eventList,
+  selectedEvent,
+}) {
+  return (
+    <EventSelector
+      onEventSelection={handleEventSelection}
+      eventList={eventList}
+      selectedEvents={[selectedEvent].filter(Boolean)}
+      showBreakBeforeButtons={false}
+      hideAllButton
+      hideClearButton={!Boolean(selectedEvent)}
+      id="event-selection"
+    />
+  );
+}
